requirements.txt
bs4
mangum
pyclean
uvicorn
pydantic
requests
python-dotenv
fastapi==0.99.0
supabase